H.W

+++ 03 - Programming Using C++ - Level 1/Input Age Variable H.W/Input Age Variable H.W.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main()
 {
-    int Age;
+    short Age;
 
     cout << "Enter your age: "; cin >> Age;
